Task A/requirements.txt
selenium==4.15.2
pandas==2.1.3
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
webdriver-manager==4.0.1
